Refactor client.rs and stdio_transport.rs
This commit refactors the code in client.rs and stdio_transport.rs.

In client.rs, the import statements for types have been modified to remove unnecessary imports. Additionally, the debug statement has been moved to the correct location.

In stdio_transport.rs, a new parameter `envs` has been added to the `new` function of `ClientStdioTransport`. This parameter allows the user to pass environment variables to the child process spawned by the transport. The `envs` parameter is a `HashMap` that maps environment variable names to their corresponding values. The `envs` parameter is then used when spawning the child process.

These changes improve the organization and functionality of the code.